pumpkin-chat/requirements.txt
flask
flask-cors
openai
reductoai
python-dotenv
requests
